refactor: split optimizer effect analysis
