+++ src/optimize/effects/builtins.rs
@@ -1,0 +1,122 @@
+pub(super) fn is_pure_non_throwing_builtin(name: &str) -> bool {
+    matches!(
+        name,
+        "strlen"
+            | "count"
+            | "intval"
+            | "floatval"
+            | "boolval"
+            | "gettype"
+            | "is_array"
+            | "is_bool"
+            | "is_float"
+            | "is_int"
+            | "is_null"
+            | "is_numeric"
+            | "is_string"
+            | "abs"
+            | "min"
+            | "max"
+            | "floor"
+            | "ceil"
+            | "round"
+            | "sqrt"
+            | "pow"
+            | "fmod"
+            | "fdiv"
+            | "sin"
+            | "cos"
+            | "tan"
+            | "asin"
+            | "acos"
+            | "atan"
+            | "atan2"
+            | "deg2rad"
+            | "rad2deg"
+            | "sinh"
+            | "cosh"
+            | "tanh"
+            | "log"
+            | "log2"
+            | "log10"
+            | "exp"
+            | "hypot"
+            | "pi"
+            | "number_format"
+            | "substr"
+            | "strpos"
+            | "strrpos"
+            | "strstr"
+            | "str_replace"
+            | "str_ireplace"
+            | "substr_replace"
+            | "strtolower"
+            | "strtoupper"
+            | "ucfirst"
+            | "lcfirst"
+            | "ucwords"
+            | "trim"
+            | "ltrim"
+            | "rtrim"
+            | "str_repeat"
+            | "strrev"
+            | "str_pad"
+            | "explode"
+            | "implode"
+            | "str_split"
+            | "strcmp"
+            | "strcasecmp"
+            | "str_contains"
+            | "str_starts_with"
+            | "str_ends_with"
+            | "ord"
+            | "chr"
+            | "nl2br"
+            | "wordwrap"
+            | "addslashes"
+            | "stripslashes"
+            | "htmlspecialchars"
+            | "htmlentities"
+            | "html_entity_decode"
+            | "urlencode"
+            | "urldecode"
+            | "rawurlencode"
+            | "rawurldecode"
+            | "md5"
+            | "sha1"
+            | "hash"
+            | "base64_encode"
+            | "base64_decode"
+            | "bin2hex"
+            | "hex2bin"
+            | "ctype_alpha"
+            | "ctype_digit"
+            | "ctype_alnum"
+            | "ctype_space"
+            | "array_key_exists"
+            | "array_search"
+            | "array_keys"
+            | "array_values"
+            | "array_merge"
+            | "array_slice"
+            | "array_combine"
+            | "array_flip"
+            | "array_reverse"
+            | "array_unique"
+            | "array_column"
+            | "array_sum"
+            | "array_product"
+            | "array_chunk"
+            | "array_pad"
+            | "array_fill"
+            | "array_fill_keys"
+            | "array_diff"
+            | "array_intersect"
+            | "array_diff_key"
+            | "array_intersect_key"
+            | "range"
+            | "json_encode"
+            | "json_decode"
+            | "json_last_error"
+    )
+}
